Create navigation.js file and move the function inside it
